+++ CHANGELOG.md
@@ -6,16 +6,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 (Date format is dd-MM-yyyy)
 
 ## Type of Changes
+
 - **Added** for new features.
 - **Changed** for changes in existing functionality.
 - **Deprecated** for soon-to-be removed features.
-- **Removed** for now removed features. 
+- **Removed** for now removed features.
 - **Fixed** for any bug fixes.
 
 [//]: ##[Unreleased]
 
+## [11.13.3] - 10-03-2022
+### Added
+- MFXTextField: added a label to specify the unit of measure (optional, leave blank string to remove)
+
+### Changed
+- Update Gradle plugins
+- Update VirtualizedFX to 11.2.5
+- Improve ROADMAP
+
 ## [11.13.2] - 09-02-2022
 ### Added
+
 - New control MFXMagnifierPane
 - ColorUtils: added some new methods to convert Colors to Strings
 - FunctionalStringConverter: added two new convenience methods